adding some PointSymbol samples.

+++ SharpMap.Providers/ArcGis/ArcGisDynamicProvider.cs
@@ -217,7 +217,7 @@ namespace SharpMap.Providers.ArcGis
 
             IFeatures features = new Features();
             IRaster raster = null;
-            IView view = new View { Resolution = resolution, Center = box.GetCentroid(), Width = (float)(box.Width / resolution), Height = (float)(box.Height / resolution) };
+            IView view = new View { Resolution = resolution, Center = box.GetCentroid(), Width = (box.Width / resolution), Height = (box.Height / resolution) };
             if (TryGetMap(view, ref raster))
             {
                 var feature = features.New();
